Introduce SigningExtension.useInMemoryPgpKeys()
This commit adds an alternative signatory provider that works with
ascii-armored in-memory PGP keys instead of keyring files. This is often
easier to setup on CI servers and more secure because there never is a
persistent file that contains the secret key. The user manual is updated
with a sample that demonstrates how to pass key and password using
environment variables.